assignment4/requirements.txt
sklearn
pandas
matplotlib
plotly
tables
notebook
gym
pymdptoolbox
mdptoolbox-hiive
tqdm
seaborn
